* INFINITE_DIST is now enum, it should cause least trouble that way. Uh, it's so hard to fight magic values these days. * Fixed crashes in battles after loading game * Fixed crash in battle AI, when stack is blocked and stands next to an enemy * Fixes problem when server's moveStack is called with dest==position * Above should cover #1053.

+++ lib/CBattleCallback.cpp
@@ -1086,7 +1086,7 @@ ReachabilityInfo CBattleInfoCallback::makeBFS(const AccessibilityInfo &accessibi
 	ret.params = params;
 
 	ret.predecessors.fill(BattleHex::INVALID);
-	ret.distances.fill(static_cast<int>(ReachabilityInfo::INFINITE_DIST));
+	ret.distances.fill(ReachabilityInfo::INFINITE_DIST);
 
 	const std::set<BattleHex> quicksands = getStoppers(params.perspective);
 	//const bool twoHexCreature = params.doubleWide;
